app/requirements.txt
numpy==1.22.3
pandas==1.4.2
requests==2.27.1
psycopg2==2.8.6
pytest==7.1.1
PyYAML==6.0
sqlalchemy==1.4.32
notebook